Test/updated_requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pyzbar==0.1.9
opencv-python==4.9.0.80
mysql-connector-python==8.4.0
python-dotenv==1.0.1
sqlalchemy>=1.4.0,<2.0
pymysql==1.1.0
requests==2.31.0
pillow==10.3.0
python-multipart==0.0.9

# For testing
qrcode[pil]==7.4.2

# Additional security and utilities
cryptography==42.0.7
validators==0.28.1